Added showing and hiding comments
